Fix test_backends
