+++ src/qibojit/tests/__init__.py
@@ -1,0 +1,6 @@
+from qibojit.custom_operators.backends import NumbaBackend, CupyBackend
+
+try:
+    backend = CupyBackend()
+except (RuntimeError, ImportError):
+    backend = NumbaBackend()
