+++ src/qibojit/tests/test_backends.py
@@ -1,28 +1,19 @@
 import pytest
 import numpy as np
-from qibo import K
-
-
-def test_engine_setter(backend_name):
-    original_backend = K.engine.name
-    K.set_engine(backend_name)
-    assert K.engine.name == backend_name
-    with pytest.raises(ValueError):
-        K.set_engine("test")
-    K.set_engine(original_backend)
+from qibojit.tests import backend as K
 
 
 @pytest.mark.parametrize("array_type", [None, "float32", "float64"])
-def test_cast(backend, array_type):
+def test_cast(array_type):
     target = np.random.random(10)
     final = K.to_numpy(K.cast(target, dtype=array_type))
     np.testing.assert_allclose(final, target)
 
 
-def test_to_numpy(backend):
+def test_to_numpy():
     x = [0, 1, 2]
     target = K.to_numpy(K.cast(x))
-    if K.engine.name == "numba":
+    if K.name == "numba":
         final = K.to_numpy(x)
     else: # pragma: no cover
         final = K.to_numpy(np.array(x))
